Small bugfix, two chars needed to be substituted.
Interesting fix, because I only caught this when someone ran a style
guide bot on my code, which suggested a non-sensical change of course.

https://github.com/larstobi/puppet-matlab/commit/42206ed0a7203ecc50422efc87048c8b7db04119

It was also interesting to have a good example of why the style guide is
"wrong" for a number of reasons. Someone should patch that instead :)

https://github.com/purpleidea/puppet-matlab/pull/1